+++ deps.ts
@@ -2,7 +2,7 @@
 // This module is browser compatible.
 export {
   AssertionError,
-} from "https://deno.land/std@0.115.1/testing/asserts.ts";
+} from "https://deno.land/std@0.118.0/testing/asserts.ts";
 export {
   isBoolean,
   isDate,
@@ -31,17 +31,20 @@ export {
 } from "https://deno.land/x/isx@v1.0.0-beta.17/mod.ts";
 
 export {
+  bgYellow,
+  black,
   blue,
+  bold,
   brightBlack,
   gray,
   green,
   red,
   yellow,
-} from "https://deno.land/std@0.116.0/fmt/colors.ts";
+} from "https://deno.land/std@0.118.0/fmt/colors.ts";
 
 // third party
 
 // deno standard module of sprintf is print error when specifier has not used
 // see https://deno.land/std@0.115.1/fmt/printf.ts#L743
 // TODO:(miayuci) re-implementation
-export { sprintf } from "https://esm.sh/sprintf-js";
+export { sprintf } from "https://cdn.skypack.dev/sprintf-js@1.1.2?dts";