+++ modifier/mod.ts
@@ -3,4 +3,6 @@
 export { not } from "./not.ts";
 export { rejects } from "./rejects.ts";
 export { resolves } from "./resolves.ts";
+export { trim } from "./trim.ts";
+export { debug } from "./debug.ts";
 export * from "./preset.ts";